Split out code from lib.rs into separate files
